Fixing Jetty example and adding code to see it run
