+++ quickstarts/web-jetty/src/main/java/org/apache/camel/cdi/jettyexample/HelloBean.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.cdi.example1;
+package org.apache.camel.cdi.jettyexample;
 
 import org.apache.camel.Body;
