--- quickstarts/web-jetty/src/main/webapp/WEB-INF/jetty-env.xml
@@ -1,16 +1,0 @@
-<!DOCTYPE Configure PUBLIC "-//Mort Bay Consulting//DTD Configure//EN" "http://www.eclipse.org/jetty/configure.dtd">
-<Configure id="webAppCtx" class="org.eclipse.jetty.webapp.WebAppContext">
-    <New id="BeanManager" class="org.eclipse.jetty.plus.jndi.Resource">
-        <Arg>
-            <Ref id="webAppCtx"/>
-        </Arg>
-        <Arg>BeanManager</Arg>
-        <Arg>
-            <New class="javax.naming.Reference">
-                <Arg>javax.enterprise.inject.spi.BeanManager</Arg>
-                <Arg>org.jboss.weld.resources.ManagerObjectFactory</Arg>
-                <Arg/>
-            </New>
-        </Arg>
-    </New>
-</Configure>
